rebase -i: Failed reword prints redundant error message
The to-do list command `reword` replays a commit like `pick` but lets
the user also edit the commit's log message. If the edited log message
is empty or is found ill-formatted by one of the commit hooks,
git-rebase--interactive prints three error messages to the console.

    1. The git-commit output, which contains all the output from hook
       scripts.
    2. A rebase diagnosis saying at which task on the to-do list it got
       stuck.
    3. Generic presumptions about what could have triggered the error.

The third message contains redundant information and does not add any
enlightenment either, which makes the output unnecessarily longish and
different from the other command's output. For instance, this is what
the output looks like if the log message is empty (contains duplicate
Signed-off-by lines).

    (1.) Aborting commit due to empty commit message. (Duplicate Signed-off-by lines.)
    (2.) Could not amend commit after successfully picking fa1afe1... Some change
    (3.) This is most likely due to an empty commit message, or the pre-commit hook
         failed. If the pre-commit hook failed, you may need to resolve the issue before
         you are able to reword the commit.

Discard the third message.

It is true that a failed hook script might not output any diagnosis but
then the generic message is not of much help either. Since this lack of
information affects the built-in git commands for commit, merge and
cherry-pick first, the solution would be to keep track of the failed
hooks in their output so that the user knows which of her hooks require
improvement.