rebase -i: reword complains about empty commit despite --keep-empty
The to-do list command `reword` replays a commit like `pick` but lets
the user also edit the commit's log message. If `--keep-empty` is passed
as option to git-rebase--interactive, empty commits ought to be replayed
without complains. However, if the users chooses to reword an empty
commit by changing the respective to-do list entry from

    pick fa1afe1 Empty commit

to

    reword fa1afe1 Empty commit

, then git-rebase--interactive suddenly fails to replay the empty
commit. This is especially counterintuitive because `reword` is thought
of as a `pick` that alters the log message in some way but nothing more
and the unchanged to-do list entry would not fail.

Handle `reword` by cherry-picking the named commit and editing the log
message using

    git commit --allow-empty --amend

instead of

    git commit --amend.

Add test.