rebase -i: reword executes pre-commit hook on interim commit
The to-do list command `reword` replays a commit like `pick` but lets
the user also edit the commit's log message. This happens in two steps.
Firstly, the named commit is cherry-picked. Secondly, the commit created
in the first step is amended using an unchanged index to edit the log
message. The pre-commit hook is meant to verify the changes introduced
by a commit (for instance, catching inserted trailing white space).
Since the committed tree is not changed between the two steps, do not
execute the pre-commit hook in the second step.

Specify the git-commit option `--no-verify` to disable the pre-commit
hook when editing the log message. Because `--no-verify` also skips the
commit-msg hook, execute the hook from within git-rebase--interactive
after the commit is created. Fortunately, the commit message is still
available in `$GIT_DIR/COMMIT_EDITMSG` after git-commit terminates.
Caveat: In case the commit-msg hook finds the new log message
ill-formatted, the user is only notified of the failed commit-msg hook
but the log message is used for the commit anyway. git-commit ought to
offer more fine-grained control over which hooks are executed.