rebase -i: Teach do_pick the option --edit
`do_pick` is the git-cherry-pick wrapper in git-rebase--interactive that
is used to implement the to-do list command `pick`. To cater for the
different pick behaviours (like `reword`), `do_pick` accepts several
options not only from the git-cherry-pick but also the git-commit
interface. Add the common option `--edit` to let the user edit the log
message of the named commit.

Loop over `$@` to parse the `do_pick` arguments. Assign the local
variable `edit` if one of the options is `--edit` so that the remainder
of `do_pick` can easily check whether the client code asked to edit the
commit message. If one of the options is unknown, mention it on the
console and `die`. Break the loop on the first non-option and do some
sanity checking to ensure that there exactly two non-options, which are
interpreted by the remainder as `<commit>` and `<title>` like before.

`do_pick` ought to act as a wrapper around `cherry-pick`. Unfortunately,
it cannot just forward `--edit` to the `cherry-pick` command line. The
assembled command line is executed within a command substitution for
controlling the verbosity of `rebase--interactive`. Passing `--edit`
would either hang the terminal or clutter the substituted command output
with control sequences. Execute the `reword` code from `do_next` instead
if the option `--edit` is specified.