rebase -i: Stop on root commits with empty log messages
The command line used to recreate root commits specifies the erroneous
option `--allow-empty-message`. If the root commit has an empty log
message, the replay of this commit shall fail and the rebase shall be
interrupted like for any other commit that is on the to-do list and has
an empty commit message. Remove the option.

The option might have been introduced by copy-and-paste of the first
part of the command line which initializes the authorship of the
sentinel commit. Indeed, the sentinel commit has an empty log message
and this should not trigger a failure, which is why the option
`--allow-empty-message` is correctly specified here.

Add test.