rebase -i: The replay of root commits is not shown with --verbose
The command line used to recreate root commits specifies the erroneous
option `-q` which suppresses the commit summary message. However,
git-rebase--interactive tends to tell the user about the commits it
creates, if she wishes (cf. command line option `--verbose`). The code
parts handling non-root commits or squash commits all output commit
summary messages. Do not make the replay of root commits an exception.
Remove the option.

It is OK to suppress the commit summary when git-commit is used to
initialize the authorship of the sentinel commit because the existence
of this additional commit is a detail of git-rebase--interactive's
implementation. The option `-q` was probably introduced as a
copy-and-paste error stemming from that part of the root commit handling
code.