rebase -i: Root commits are replayed with an unnecessary option
The command line used to recreate root commits specifies the effectless
option `-C`. It is used to reuse commit message and authorship from the
named commit but the commit being amended here, which is the sentinel
commit, already carries the authorship and log message of the processed
commit. Note that the committer email and commit date fields do not
match the root commit either way. Remove the option. However, `-C`
(other than `-c`) does not invoke the editor and the `--amend` option
invokes it by default. Disable editor invocation again by specifying
`--no-edit`.