rebase -i: Teach do_pick the option --amend
`do_pick` is the git-cherry-pick wrapper in git-rebase--interactive that
is used to implement the to-do list commands `pick`, `reword` and
`edit`. To cater for the different pick behaviours (like `squash`),
`do_pick` accepts several options not only from the git-cherry-pick but
also the git-commit interface.

Add the option `--amend` from the git-commit interface to the options
pool of `do_pick`. It creates a new commit for the changes introduced by
the picked commit and the previous one. The previous commit is then
replaced with the new commit. If no other options are specified, the log
message of the previous commit is used.

Be careful when `--amend` is used to pick a root commit because HEAD
might point to the sentinel commit but there is still nothing to amend.
Be sure to initialize `amend` so that commits are squashed even when
git-rebase--interactive is interrupted for resolving conflicts. It is
not a mistake to do the initialization regardless of any conflicts
because `amend` is always cleared before the next to-do item is
processed.