rebase -i: Teach do_pick the option --file
`do_pick` is the git-cherry-pick wrapper in git-rebase--interactive that
is used to implement the to-do list command `pick`, `reword` and `edit`.
To cater for the different pick behaviours (like `squash`), `do_pick`
accepts several options not only from the git-cherry-pick but also the
git-commit interface.

Add the option `--file` from the git-commit interface to the options
pool of `do_pick`. It expects an argument itself which is interpreted as
a file path and takes the commit message from the given file. If
`--file` is passed to `do_pick`, assign the given file path to the local
variable `rewrite_message` and relay the option

    --file "$rewrite_message"

to the git-commit command line which creates the commit.