rebase -i: Implement squash in terms of do_pick
The to-do list command `squash` and its close relative `fixup` replay
the changes of a commit like `pick` but do not recreate the commit.
Instead they replace the previous commit with a new commit that also
introduces the changes of the squashed commit. This is roughly like
cherry-picking without committing and using git-commit to amend the
previous commit.

The to-do list

    pick   a Some changes
    squash b Some more changes

gets translated into the sequence of git commands

    git cherry-pick a
    git cherry-pick -n b
    git commit --amend

and if git-cherry-pick supported `--amend` this would look even more
like the to-do list it is based on

    git cherry-pick a
    git cherry-pick --amend b.

Since `do_pick` takes care of `pick` entries and the above suggests
`squash` as an alias for `pick --amend`, reimplement `squash` in terms
of `do_pick --amend`. Introduce `$squash_msg` as the commit message via
the `--file` option. When the last commit of a squash series is
processed, the user is asked to review the log message. Pass `--edit` as
additional option in this case. The only difference in the options
passed to git-commit and `do_pick` is the omitted `--no-verify`.
However, `do_pick` does not execute the verification hooks anyway
because it solely replays commits and assumes that they have been
verified before.