rebase -i: Teach do_pick the option --reset-author
`do_pick` is the git-cherry-pick wrapper in git-rebase--interactive that
is used to implement many of the to-do list commands. Eventually, the
complete `do_pick` interface will be exposed to the user in some form or
another and those commands will become simple aliases for the `do_pick`
options now used to implement them.

Add the git-commit option `--reset-author` to the options pool of
`do_pick`. It rewrites the author date and name of the picked commit to
match the committer date and name.

If `--reset-author` is passed to `do_pick`, set the `rewrite` flag and
relay the option to the git-commit command line which creates the final
commit. If `--amend` is not passed as well, the fresh authorship effect
is achieved by the mere fact that we are creating a new commit.