rebase -i: Teach do_pick the option --signoff
`do_pick` is the git-cherry-pick wrapper in git-rebase--interactive that
is currently used to implement most of the to-do list commands and
offers additional options that will eventually find their way onto to-do
lists.

To extend the repertoire of available options, add the git-commit and
git-cherry-pick option `--signoff` to the `do_pick` interface. It
appends a Signed-off-by: line using the committer identity to the log
message of the picked commit.