rebase -i: Parse to-do list command line options
Read in to-do list lines as

    command args

instead of

    command sha1 rest

so that to-do list command lines can specify additional arguments apart
from the commit hash and the log message title, which become the
non-options in `args`. Loop over `args`, put all options (an argument
beginning with a dash) in `opts`, stop the loop on the first non-option
and assign it to `sha1`. The loop does not know the options it parses so
that options that take an argument themselves are not supported at the
moment. Neither are options that contain spaces because the shell
expansion of `args` in `do_next` interprets white space characters as
argument separator, that is a command line like

    pick --author "A U Thor" fa1afe1 Some change

is parsed as the pick command

    pick --author

and the commit hash

    "A

which obviously results in an unknown revision error. For the sake of
completeness, in the example above the message title variable `rest` is
assigned the string 'U Thor" fa1afe1 Some change' (without the single
quotes).

Print an error message for unknown or unsupported command line options,
which means an error for all specified options at the moment. Cleanly
break the `do_next` loop by assigning the special value 'unknown' to the
local variable `command`, which triggers the unknown command case in
`do_cmd`.

The to-do list is also parsed when the commit hashes are translated
between long and short format before and after the to-do list is edited.
Apply the same procedure as in `do_cmd` with the exception that we only
care about where the options stop and the commit hash begins. Do not
reject any options when transforming the commit hashes.