rebase -i: Enable options `--signoff`, `--reset-author` for pick, reword
pick and reword are atomic to-do list commands in the sense that they
open a new task which is closed after the respective command is
completed. squash and fixup are not atomic. They create a new task which
is not completed until the last squash or fixup is processed.

Lift the general unknown option blockade for the pick and reword
commands. If `do_cmd` comes across one of the options `--signoff` and
`--reset-author` while parsing a to-do entry and the scheduled command
is either `pick` or `reword`, relay the option to `do_pick`.

The `do_pick` options `--gpg-sign` and `--file` are not yet supported
because `do_cmd` cannot handle option arguments and options with spaces
at the moment. It is true that edit is one of the atomic commands but it
displays hash information when the rebase is stopped and some options
rewrite the picked commit which alters that information. squash and
fixup still do not accept user options as the interplay of
`--reset-author` and the author script are yet to be determined.